Transaction category deleting modal
