--- src/app/common/transaction-category-deleting-modal/transaction-category-deleting-modal.component.html
@@ -1,20 +1,0 @@
-<h1 mat-dialog-title>{{data.headerText}}</h1>
-<div mat-dialog-content>{{data.confirmationText}}
-
-</div>
-<form [formGroup]="form" class="form__container" *ngIf="!isLoadingTransactionCategories">
-        <mat-form-field>
-        <mat-label>Select category (optional)</mat-label>
-                <mat-select [formControl]="form.controls['category']">
-                <mat-option *ngFor="let category of transactionsCategories" [value]="category.name">
-                {{ category.name }}
-                </mat-option>
-                </mat-select>
-        </mat-form-field>
-</form>
-<div mat-dialog-actions>
-    <button mat-stroked-button color="primary"
-            [mat-dialog-close]="false">{{ data.denyBtnText || 'No' }}</button>
-    <button mat-raised-button color="primary"
-            [mat-dialog-close]="true">{{ data.confirmBtnText || 'Yes' }}</button>
-</div>
