+++ src/app/pages/categories/management/categories-deleting/transaction-category-deleting-modal.component.html
@@ -1,0 +1,20 @@
+<mat-spinner class="loading-container" *ngIf="isLoadingTransactionCategories" [diameter]="30"></mat-spinner>
+<div class="modal-width" *ngIf="!isLoadingTransactionCategories">
+        <h1 mat-dialog-title>{{data.headerText}}</h1>
+        <div mat-dialog-content class="confirm-text">{{data.confirmationText}}</div>
+        <mat-form-field>
+                <mat-label>Select category (optional)</mat-label>
+                <mat-select [formControl]="form">
+                        <mat-option *ngFor="let category of transactionsCategories" [value]="category.name">
+                                {{ category.name }}
+                        </mat-option>
+                </mat-select>
+        </mat-form-field>
+        <div mat-dialog-actions>
+        <button mat-stroked-button color="primary"
+                [mat-dialog-close]="false">{{ data.denyBtnText || 'No' }}</button>
+        <button mat-raised-button color="primary"
+                [mat-dialog-close]="true">{{ data.confirmBtnText || 'Yes' }}</button>
+        </div>
+</div>
+
