+++ src/app/pages/categories/management/categories-deleting/transaction-category-deleting-modal.component.scss
@@ -1,0 +1,18 @@
+.confirm-text {
+    text-align: center;
+    padding-bottom: 10px;
+}
+
+.form__container {
+    display: flex;
+    flex-direction: column;
+    gap: 5px;
+  }
+
+.loading-container {
+    margin: auto;
+}
+
+.modal-width {
+    width: 500px;
+}
