+++ src/app/pages/categories/management/categories-deleting/transaction-category-deleting-modal.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { Observable, map } from 'rxjs';
+import { Observable, map, tap } from 'rxjs';
 import { TransactionCategoryDeletingModalComponent } from './transaction-category-deleting-modal.component';
 import { ITransactionCategoryDeletingModalData } from './transaction-category-deleting-modal.types';
